Tests

+++ src/main/java/com/example/bandproject/demo/services/BandMembershipService.java
@@ -7,6 +7,7 @@ import com.example.bandproject.demo.models.User;
 import com.example.bandproject.demo.repositories.BandMembershipRepository;
 import com.example.bandproject.demo.repositories.BandRepository;
 import com.example.bandproject.demo.repositories.UserRepository;
+import javassist.NotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -24,19 +25,29 @@ public class BandMembershipService {
         this.bandRepository = bandRepository;
     }
 
-    public void newMembership(AddMembership addMembership){
-        User user = userRepository.findById(addMembership.getMemberId()).get();
-        Band band = bandRepository.findById(addMembership.getBandId()).get();
+    public BandMembership newMembership(AddMembership addMembership) throws NotFoundException {
+        if (userRepository.findById(addMembership.getMemberId()).isPresent() && bandRepository.findById(addMembership.getBandId()).isPresent()) {
+            User user = userRepository.findById(addMembership.getMemberId()).get();
+            Band band = bandRepository.findById(addMembership.getBandId()).get();
 
-        BandMembership bandMembership = new BandMembership(user, band);
+            BandMembership bandMembership = new BandMembership(user, band);
 
-        bandMembershipRepository.save(bandMembership);
+           return bandMembershipRepository.save(bandMembership);
+        } else {
+            throw new NotFoundException(" Band or User Not found");
+        }
     }
-    public void deleteMembership(Long membershipId){
-        bandMembershipRepository.deleteById(membershipId);
+
+    public void deleteMembership(Long membershipId) throws NotFoundException {
+
+        if (bandMembershipRepository.findById(membershipId).isPresent()) {
+            bandMembershipRepository.deleteById(membershipId);
+        } else {
+            throw new NotFoundException("this membership does not exist");
+        }
     }
 
-    public List<BandMembership> getAllMemberships(){
+    public List<BandMembership> getAllMemberships() {
         return bandMembershipRepository.findAll();
     }
 }
